+++ R/TROLLv3_species.R
@@ -8,6 +8,6 @@
 #'   \item{s_wsg}{wood specific gravity} \item{s_dbhmax}{maximum diameter}
 #'   \item{s_hmax}{maximum height} \item{s_ah}{height-diameter allometry
 #'   coefficient} \item{s_regionalfreq}{regional frequency} \item{s_tlp}{Turgor
-#'   loss point} \item{s_drymass}{leaf dry mass} }
-#'
+#'   loss point} \item{s_drymass}{leaf dry mass} \item{s_seedmass}{seed mass} }
+#'   
 "TROLLv3_species"